+++ Modelos/LibroPedidosDetalleDto.cs
@@ -16,7 +16,7 @@ namespace Modelos
         public string StripeSessionId { get; set; }
 
         [Required]
-        public long CostoTotal { get; set; }
+        public double CostoTotal { get; set; }
 
         [Required]
         public int LibroId { get; set; }
